requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
pillow==11.0.0
ultralytics==8.3.28
ultralytics-thop==2.0.9
paddleocr==2.9.1
paddlepaddle==2.6.1